feat: Build api methods and shared action creator
Created a new file to create simplified api calling functions.
Added getInitialData() which internally makes two API calls to get users and question from database.
Added two more methods to save question and answer to the backend.
Only this file will be exposed to the frontend componnets and not the Data.js file.

Created new thunked action creator handleInitialData() which makes call to the API to get data.
Then uses that data to update the state of store using dispatch().

It is not yet being used anywhere