feat: Add redux store to application
Created new redux store by using createStore() of redux and passed root reducer and middlewares to it.

Surrounded the App component with Provider component provided by 'react-redux' and passed store as prop to it. It will enabled application now to brodcast store to all components which menas we will not have to send data as props to nestedeeply nested components.

Components at the moment are not using this store