feat: Apply LoadingBar with 'react-redux-loading' when initialData is loaded
Added a loading bar at the top of page when is shown when initialData is being fetched from API and saved to store.
Passed the loadingBar reducer to the root reducer.
Utilised the provided showLoading() and hideLoading() action creator of package to show loading bar when thunked action creator is called
and hide loadingBar once both API and dispatch operations are complete.

Used LoadingBar UI component provided by the package to render the loading bar on page