fix: Show selected answer on PollDetails when user submit answer to question
Poll details were not ighlighting the user selected answer when user submitted the answer for a question on question page.
Fixed this issue by passing the selected option from the component state because previously it was being passed from props of Poll Component to PollDetails.
Change in component state now re-renders the view