feat: Add new action creator to Save the answer to user profile
It was missed previously when save answer was implemented.
Previous implementation only included saving question's answer to questions state but
it should of also added the id of answered question to user.answers object along with option choosen.

Creator new action type and creator to handle this operation.
Also added one more case to update the users state in the store correctly