fix: Show only answered/unanswered questions using tabs
Tabs were implemented at the beginning of project but both tab panels were visible at all times.
This was due to the issue of 'jquery' in 'materlize-library'.
Fixed it by initialising the Materialize Tab in componentDidMount() lifecycle hook of react component.

Now user is able to toggle betwwen answered and unanswered question